connect managers to socket services

+++ library/core/src/main/java/com/nextfeed/library/core/service/manager/dto/mood/NewMoodRequest.java
@@ -10,6 +10,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class NewMoodRequest {
-    private Integer moodValue;
+    private Double moodValue;
     private Integer participantsCount;
 }
